move tests and code to solaredge
